+++ filters.php
@@ -1,6 +1,7 @@
 <?php
 
 use Admin\Libraries\ModelConfig;
+use Admin\Libraries\SettingsConfig;
 
 //Filters
 
@@ -14,7 +15,7 @@ Route::filter('add_assets', function()
 
 	$assets->add('jquery.ui.css', 'css/ui/jquery-ui-1.9.1.custom.min.css');
 	$assets->add('jquery.ui.timepicker.css', 'css/ui/jquery.ui.timepicker.css');
-	$assets->add('chosen_css', 'css/chosen.css');
+	$assets->add('select2_css', 'js/jquery/select2/select2.css');
 	$assets->add('jquery.lw-colorpicker', 'css/jquery.lw-colorpicker.css');
 	$assets->add('main_style', 'css/main.css');
 
@@ -26,12 +27,11 @@ Route::filter('add_assets', function()
 	//jquery core
 	$assets->add('jquery', 'js/jquery/jquery-1.8.2.min.js');
 
-	//jquery chosen and ajax chosen
-	$assets->add('jquery-chosen', 'js/jquery/jquery.chosen.min.js');
-	$assets->add('jquery-ajax-chosen', 'js/jquery/jquery.ajax-chosen.min.js');
+	//select2
+	$assets->add('jquery-select2', 'js/jquery/select2/select2.js');
 
 	//jquery ui
-	$assets->add('jquery.ui', 'js/jquery/jquery-ui-1.9.1.custom.min.js');
+	$assets->add('jquery.ui', 'js/jquery/jquery-ui-1.10.2.custom.min.js');
 
 	//jquery timepicker addon
 	$assets->add('jquery.ui.timepicker', 'js/jquery/jquery-ui-timepicker-addon.js');
@@ -77,6 +77,7 @@ Route::filter('add_assets', function()
 	//and finally the site scripts
 	$assets->add('page', 'js/page.js');
 	$assets->add('admin', 'js/admin.js');
+	$assets->add('settings', 'js/settings.js');
 });
 
 
@@ -110,6 +111,21 @@ Route::filter('validate_model', function ()
 {
 	$modelName = Request::route()->parameters[0];
 	$config = ModelConfig::get($modelName);
+
+	//if the model doesn't exist at all, redirect to 404
+	if (!$config)
+	{
+		return Response::error('404');
+	}
+
+	Request::route()->parameters[0] = $config;
+});
+
+//validate_settings filter
+Route::filter('validate_settings', function ()
+{
+	$settingsName = Request::route()->parameters[0];
+	$config = SettingsConfig::get(SettingsConfig::$prefix . $settingsName);
 
 	//if the model doesn't exist at all, redirect to 404
 	if (!$config)