+++ language/eu/knockout.php
@@ -1,0 +1,24 @@
+<?php 
+
+return array(
+
+	/*
+	|--------------------------------------------------------------------------
+	| Administrator Language Lines for knockout
+	|--------------------------------------------------------------------------
+	|
+	| eu - Basque
+	|
+	*/
+
+	'delete_active_item' => 'Ziur zaude item hau borratu nahi duzula? Ekintza hau itzulezina da.',
+	'saving' => 'Gordetzen...',
+	'saved' => 'Item-a gorde da.',
+	'deleting' => 'Ezabatzen...',
+	'deleted' => 'Item-a ezabatu da.',
+	'character_left' => ' karaktere faltan',
+	'characters_left' => ' karaktere faltan',
+	'no_results' => 'Emaitzik ez',
+	'select_options' => 'Hautatu Aukera Batzuk',
+
+);