+++ language/fr/knockout.php
@@ -1,0 +1,24 @@
+<?php 
+
+return array(
+
+	/*
+	|--------------------------------------------------------------------------
+	| Administrator Language Lines for knockout
+	|--------------------------------------------------------------------------
+	|
+	| fr - Français
+	|
+	*/
+
+	'delete_active_item' => 'Etes-vous certain de vouloir supprimer cet élément ?L\'opération sera irréversible.',
+	'saving' => 'Sauve...',
+	'saved' => 'Elément sauvegardé.',
+	'deleting' => 'Suppression...',
+	'deleted' => 'Elément supprimé.',
+	'character_left' => ' caractère restant',
+	'characters_left' => ' caractères restants',
+	'no_results' => 'Aucun résultat ne correspond',
+	'select_options' => 'Sélectionner des options',
+
+);